patata

+++ source_code/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: fmontero <fmontero@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/10 23:30:38 by fmontero          #+#    #+#             */
-/*   Updated: 2024/04/10 23:30:40 by fmontero         ###   ########.fr       */
+/*   Updated: 2024/04/11 11:04:33 by fmontero         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	size_t	len;
-	size_t	i;
-	char	*result;
+	unsigned int	len;
+	size_t			i;
+	char			*result;
 
-	len = ft_strlen(s);
+	len = (unsigned int)ft_strlen(s);
 	result = (char *)malloc(len + 1);
 	if (result == NULL)
 		return (NULL);
@@ -27,7 +27,7 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 	i = 0;
 	while (i < len)
 	{
-		result[i] = f((unsigned int)i, s[i]);
+		result[i] = f(i, s[i]);
 		i++;
 	}
 	return (result);
